024.swap-nodes-in-pairs/problem24_test.go
// Given a linked list, swap every two adjacent nodes and return its head.

// For example,
// Given 1->2->3->4, you should return the list as 2->1->4->3.

// Your algorithm should use only constant space. You may not modify the values in the list, only nodes itself can be changed.

package leetcode

import (
	"reflect"
	"testing"
)

func Test_swapPairs(t *testing.T) {
	type args struct {
		head *ListNode
	}
	tests := []struct {
		name string
		args args
		want *ListNode
	}{
		{
			"[Test Case 1]",
			args{
				&ListNode{
					1,
					&ListNode{
						2,
						nil,
					},
				},
			},
			&ListNode{
				2,
				&ListNode{
					1,
					nil,
				},
			},
		},
		{
			"[Test Case 2]",
			args{
				nil,
			},
			nil,
		},
		{
			"[Test Case 3]",
			args{
				&ListNode{
					1,
					&ListNode{
						2,
						&ListNode{
							3,
							&ListNode{
								4,
								nil,
							},
						},
					},
				},
			},
			&ListNode{
				2,
				&ListNode{
					1,
					&ListNode{
						4,
						&ListNode{
							3,
							nil,
						},
					},
				},
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := swapPairs(tt.args.head); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("swapPairs() = %v, want %v", got, tt.want)
			}
		})
	}
}
